requirements.txt
mcp
fastmcp
httpx
python-dotenv
appwrite
async-timeout
pydantic
starlette
uvicorn
